Update and rename HW7 - S23 - solution.R to Classification Tree.R
A classification Tree Model is developed to predict the severity of automobile accidents. This model aims to classify accidents into categories based on whether they involve an injury ("fatal" or "non-fatal") or do not involve an injury ("no-injury"). The features used for classification include rush hour status, work zone presence, whether the accident occurred on a weekday or weekend, if it happened on an interstate highway, light conditions during the accident, the road level, speed limit, surface condition, type of traffic flow, and weather conditions.

The process involves several key steps:

Building the Classification Tree Model: A model is trained using the rpart function in R, considering all the aforementioned features to predict the injury outcome of accidents.
Visualization: The structure of the classification tree is visualized using the prp function, which helps in understanding how decisions are made within the model.
Model Application: The trained model is then applied to both the training and validation datasets to predict the injury status of accidents.
Performance Evaluation: The model's accuracy and specificity are assessed through a confusion matrix, providing insights into its effectiveness in correctly classifying accidents.
Real-world Application Test: Finally, the model is tested with a new dataset to predict the injury status of a hypothetical traffic accident, demonstrating its practical applicability.